CBP - stock Transactions - Crypto Money Restock - QA Fix  GetManager #3135
